// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "productA".
 */
export interface ProductA {
  id: number;
  title: string;
  /**
   * URL-friendly version of the title
   */
  slug?: string | null;
  description: string;
  price: number;
  productAImage: number | Media;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}